+++ test/unit/data.js
@@ -40,6 +40,7 @@
     ok( is('Number', $reel.data('opening_ticks')), '`opening_ticks` Number');
     ok( is('Object', $reel.data('options')), '`options` Object');
     ok( is('Boolean', $reel.data('playing')), '`playing` Boolean');
+    ok( is('Number', $reel.data('preloaded')), '`preloaded` Number');
     ok( is('Boolean', $reel.data('reeling')), '`reeling` Boolean');
     ok( is('Number', $reel.data('revolution')), '`revolution` Number');
     ok( is('Number', $reel.data('row')), '`row` Number');