+++ CHANGELOG.markdown
@@ -33,6 +33,7 @@ Development
   failsafe cursor is now `"move"`.
 * New `sequence` option for easy definition of sequence frames by one
   path with a counter placeholder.
+* Default height of the preloader progress bar reduced to 2 pixels.
 
 
 Version 1.1.3